requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart
python-dotenv
minio
pymysql
PyJWT
cryptography
httpx
strawberry-graphql
pytest
pytest-asyncio